requirements.txt
django>=4.2
# NumPy 2.x breaks TensorFlow 2.13 and other wheels built for 1.x (pulled in via langchain → transformers).
numpy>=1.24,<2
pandas>=2.0
scikit-learn>=1.3
joblib>=1.3
scipy>=1.11

# ASGI + ALIA RAG (FastAPI, mounted in alia.asgi)
uvicorn[standard]>=0.27
starlette>=0.37
fastapi>=0.110
pydantic>=2.0
openai-whisper>=20231117
pyttsx3>=2.90

# LangChain stack (Ollama + Chroma; requires local Ollama)
langchain-core>=0.3
langchain-chroma>=0.1
langchain-ollama>=0.2
langchain-community>=0.3
chromadb>=0.5
rank-bm25>=0.2
google-genai>=1.14.0
python-dotenv>=1.0.1
